strict care formatting size_t // rm(trailing white space)
git-svn-id: https://svn.r-project.org/R/trunk@72926 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -4106,7 +4106,7 @@ attribute_hidden size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize
     while ((c = Rconn_fgetc(con)) != R_EOF)
     {
         if (nbuf + 1 >= bufsize)
-            error(_("line longer than buffer size %d"), bufsize);
+            error(_("line longer than buffer size %lu"), (unsigned long)bufsize);
         if (c != '\n')
         {
             buf[++nbuf] = (char)c;
@@ -4123,7 +4123,7 @@ attribute_hidden size_t Rconn_getline(Rconnection con, char *buf, size_t bufsize
     if (nbuf >= 0 && buf[nbuf])
     {
         if (nbuf + 1 >= bufsize)
-            error(_("line longer than buffer size %d"), bufsize);
+            error(_("line longer than buffer size %lu"), (unsigned long)bufsize);
         buf[++nbuf] = '\0';
     }
     return (nbuf);
